moves generateToken to middlewares
