+++ config/middlewares.js
@@ -5,7 +5,20 @@ const jwtKey = require('../_secrets/keys').jwtKey;
 // quickly see what this file exports
 module.exports = {
   authenticate,
+  generateToken
 };
+
+function generateToken(user) {
+  const payload = {
+    subject: user.id,
+    username: user.username
+  };
+  const secret = jwtKey;
+  const options = {
+    expiresIn: '15m'
+  };
+  return jwt.sign(payload, secret, options);
+}
 
 // implementation details
 function authenticate(req, res, next) {
@@ -21,7 +34,7 @@ function authenticate(req, res, next) {
     });
   } else {
     return res.status(401).json({
-      error: 'No token provided, must be set on the Authorization Header',
+      error: 'No token provided, must be set on the Authorization Header'
     });
   }
 }
